# Some CBot code optimizations