--- a/src/CBot/CBotVar/CBotVar.cpp
+++ b/src/CBot/CBotVar/CBotVar.cpp
@@ -49,7 +49,7 @@ namespace CBot
 long CBotVar::m_identcpt = 0;
 
 ////////////////////////////////////////////////////////////////////////////////
-CBotVar::CBotVar( )
+CBotVar::CBotVar( ) : m_token(nullptr)
 {
     m_pMyThis = nullptr;
     m_pUserPtr = nullptr;
@@ -62,9 +62,17 @@ CBotVar::CBotVar( )
     m_mPrivate = ProtectionLevel::Public;
 }
 
-CBotVar::CBotVar(const CBotToken &name) : CBotVar()
+CBotVar::CBotVar(const CBotToken &name) : m_token(new CBotToken(name))
 {
-    m_token = new CBotToken(name);
+    m_pMyThis = nullptr;
+    m_pUserPtr = nullptr;
+    m_InitExpr = nullptr;
+    m_LimExpr = nullptr;
+    m_type  = -1;
+    m_binit = InitType::UNDEF;
+    m_ident = 0;
+    m_bStatic = false;
+    m_mPrivate = ProtectionLevel::Public;
 }
 
 ////////////////////////////////////////////////////////////////////////////////